Adeed sweet alert and dark mode

+++ src/pages/MyServices.jsx
@@ -2,6 +2,7 @@ import React, { useContext, useEffect, useState } from 'react';
 import { AuthContext } from '../provider/AuthProvider';
 import { Link } from 'react-router';
 import axios from 'axios';
+import Swal from 'sweetalert2';
 
 const MyServices = () => {
 
@@ -21,23 +22,48 @@ const MyServices = () => {
     console.log(myServices);
 
     const handleDelete = (id) => {
-        axios.delete(`http://localhost:3000/delete/${id}`)
-        .then(res => {
-            console.log(res.data);
 
-            
-            const filterData = myServices.filter(service => service._id != id)
 
-            setMyServices(filterData)
-            
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
 
-        })
-        .catch(err=> {
-            console.log(err);
-            
-        })
+                axios.delete(`http://localhost:3000/delete/${id}`)
+                    .then(res => {
+                        console.log(res.data);
+                        if (res.data.deletedCount == 1) {
+
+                            const filterData = myServices.filter(service => service._id != id)
+
+                            setMyServices(filterData)
+                            
+                            Swal.fire({
+                                title: "Deleted!",
+                                text: "Your file has been deleted.",
+                                icon: "success"
+                            });
+                        }
+
+                    })
+                    .catch(err => {
+                        console.log(err);
+
+                    })
+
+
+            }
+        });
+
+
     }
-    
+
 
 
 
@@ -82,7 +108,7 @@ const MyServices = () => {
                                     </td>
                                     <td>{service?.price}</td>
                                     <th className='flex gap-3'>
-                                        <button onClick={()=>handleDelete(service?._id)} className="btn btn-error btn-xs">Delete</button>
+                                        <button onClick={() => handleDelete(service?._id)} className="btn btn-error btn-xs">Delete</button>
                                         <Link to={`/update-service/${service?._id}`}><button className="btn btn-primary btn-xs">Edit</button></Link>
                                     </th>
                                 </tr>
